refactor: 컴포넌트 디렉터리 정리
