+++ src/components/index.js
@@ -1,5 +1,6 @@
 import Footer from "./Footer";
 import Header from "./Header";
-import PrivateRoute from "./PrivateRoute"
+import PrivateRoute from "./PrivateRoute";
+import LoadingIndicator from "./LoadingIndicator";
 
-export { Footer, Header, PrivateRoute };
+export { Footer, Header, PrivateRoute, LoadingIndicator };
