--- src/components/pagination.css
@@ -1,56 +1,0 @@
-.pagination {
-    display: flex;
-    justify-content: center;
-    margin-top: 15px;
-  }
- 
-  ul {
-    list-style: none;
-    padding: 0;
-  }
- 
-  ul.pagination li {
-    display: inline-block;
-    width: 30px;
-    height: 30px;
-    margin: 3px;
-    border: 1px solid #e2e2e2;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1rem;
-    border-radius: 50%;
-  }
- 
-  ul.pagination li:first-child{
-    border-radius: 50%;
-  }
- 
-  ul.pagination li:last-child{
-    border-radius: 50%;
-  }
- 
-  ul.pagination li a {
-    text-decoration: none;
-    color: black;
-    font-size: 10px;
-  }
- 
-  ul.pagination li.active a {
-    color: white;
-  }
- 
-  ul.pagination li.active {
-    background-color: gray;
-  }
- 
-  ul.pagination li a:hover,
-  ul.pagination li a.active {
-    color: gray;
-  }
- 
-  .page-selection {
-    width: 48px;
-    height: 30px;
-    color: #337ab7;
-  }
