+++ src/components/pagination/Pagination.scss
