+++ src/pages/gallery/GalleryWrite.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useEffect } from "react";
 import { useNavigate } from 'react-router-dom';
-import QuillEditor from '../../components/QuillEditor';
+import QuillEditor from '../../components/editor/QuillEditor';
 
 function GalleryWrite() {
     const navigate = useNavigate();
