+++ src/pages/tuxinfo/Tuxinfo02.jsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import './style.css'
 import { Link } from 'react-router-dom';
 
 const HistoryBox = ({ props }) => {
