+++ src/style.css
@@ -1,3 +1,7 @@
+/*
+    Header.jsx, Footer.jsx, Tuxinfo02.jsx에서 사용하는 CSS
+*/
+
 .nav a {
     padding: 27.5px 20px;
 }
